requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
